requirements.txt
# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
selenium==4.16.0
webdriver-manager==4.0.1

# Data Processing
pandas==2.2.0
python-dateutil==2.8.2

# OCR and Document Processing
docling==1.0.0

# Vector Database
faiss-cpu==1.7.4
numpy==1.26.3

# MCP Server
fastmcp==0.1.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
sse-starlette==2.0.0

# LLM Integration
langchain==0.1.5
langchain-community==0.0.16
tiktoken==0.5.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0